+++ include/shared/State.h
@@ -91,6 +91,7 @@ extern POINT g_dragStartHUD;
 extern POINT g_dragStartMouse;
 extern HWND g_hHUD;
 extern HWND g_hPanel;
+extern bool g_pendingShowHUD;
 extern std::atomic<int> g_loadingProgress;
 
 void __cdecl RefreshHotkeys(HWND hWnd);